# api/swagger/swagger.json
{
  "info": {
    "title": "Tenant Service",
    "description": "This service is responsible for segmenting service configuration between distinct client applications",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
...
    "/accounts": {
      "post": {
        "summary": "Save Account",
        "description": "Save Account",
        "operationId": "saveAccount",
        "tags": [
          "accounts"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "description": "account",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Account"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "Account",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update Account",
        "description": "Update Account",
        "operationId": "updateAccount",
        "tags": [
          "accounts"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "account",
            "description": "account",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Account"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/accounts/{id}": {
      "get": {
        "summary": "Get Account By Id",
        "description": "Get Account By Id",
        "operationId": "getAccount",
        "tags": [
          "accounts"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of Account",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Account"
            }
          },
          "404": {
            "description": "Not found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/apps": {
      "post": {
        "summary": "Save Application",
        "description": "Save Application",
        "operationId": "saveApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "application",
            "description": "application",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "201": {
            "description": "Application",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "put": {
        "summary": "Update Application",
        "description": "Update Application",
        "operationId": "updateApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "application",
            "description": "application",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            },
            "in": "body",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Application",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "400": {
            "description": "Bad request",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "get": {
        "summary": "Get Applications",
        "description": "Get Page of Application(s)",
        "operationId": "getApplications",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "search",
            "description": "Text Search String",
            "type": "string",
            "in": "query",
            "required": false
          },
          {
            "name": "page",
            "description": "Page Number ( 0 index based )",
            "type": "integer",
            "format": "int64",
            "required": false
          },
          {
            "name": "size",
            "description": "Page Size - default 10",
            "type": "integer",
            "format": "int64",
            "required": false
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "PageResponse",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          }
        }
      }
    },
    "/apps/{id}": {
      "get": {
        "summary": "Get Application By Id",
        "description": "Get Application By Id",
        "operationId": "getApplication",
        "tags": [
          "applications"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Id of Application",
            "type": "string",
            "in": "path",
            "required": true
          },
          {
            "name": "x-fast-pass",
            "description": "Bypass Auth",
            "type": "boolean",
            "in": "header",
            "require": false
          }
        ],
        "responses": {
          "200": {
            "description": "Application",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Application"
            }
          },
          "404": {
            "description": "Not found",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        },
        "violations": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ConstraintViolation"
          }
        }
      }
    },
    "ConstraintViolation": {
      "type": "object",
      "required": [
        "param",
        "message"
      ],
      "properties": {
        "param": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
...
    "Account": {
      "type": "object",
      "required": [
        "accountNumber",
        "tenantId"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "accountNumber": {
          "type": "string"
        },
        "tenantId": {
          "type": "string"
        }
      }
    }
  },
...
  "securityDefinitions": {},
...
}